+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+
+- tezos.tzkt: Fixed an issue with approving schema after reindexing triggered by rollback.
+- evm.subsquid: Last mile indexing is significantly faster now.
+
 ## [7.2.1] - 2023-12-12
 
 ### Added